+++ src/main/java/com/livares/intern/entity/Product.java
@@ -5,7 +5,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -18,7 +17,7 @@ import lombok.Setter;
 public class Product {
  
 	@Id
-	int id;
+	long id;
 	String name;
 	String description;
 	String img;
