+++ src/main/java/com/livares/intern/entity/Users.java
@@ -1,6 +1,5 @@
 package com.livares.intern.entity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -13,12 +12,13 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Users {
+public class Users  {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-     int id;
+     long id;
 	String firstName;
 	String lastName;
      String username;
      String password;
+     
 }
