+++ src/main/java/com/livares/intern/repository/CategoryRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.livares.intern.entity.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer>{
+public interface CategoryRepository extends JpaRepository<Category, Long>{
 
+	
 }
